Split large storage and replication tests
