+++ crates/truthdb-core/src/engine/tests/replication/mod.rs
@@ -1,0 +1,10 @@
+use crate::engine::test_support::*;
+
+use crate::engine::*;
+
+mod readable_standby;
+mod restartpoints;
+mod slots;
+mod synchronous_commit;
+mod timelines;
+mod transport;
